[#fix] 장비 벨류 표기 오류 수정
amount를 계산할 때 소수로 하지 않아 0으로 내림됨

+++ Assets/Scenes/General/Scripts/StaticDatas/Scripts/StaticEquipmentValueData.cs
@@ -56,7 +56,7 @@ public class EquipmentValueData
     {
         float minValue = min_value[Array.FindIndex(buff_type, item => item == buff_type[index])];
         float maxValue = max_value[Array.FindIndex(buff_type, item => item == buff_type[index])];
-        float amount = (level - min_level) / (max_level - min_level);
+        float amount = (float)(level - min_level) / (float)(max_level - min_level);
         float value = (amount * (maxValue - minValue)) + minValue;
         return value;
     }
